Connections between nodes
+ Various bug fixes

+++ Assets/If Simulator/Scripts/AI/Behavior Tree/Editor/BTreeView.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using UnityEditor;
 using UnityEditor.Experimental.GraphView;
@@ -16,6 +17,8 @@ namespace BehaviorTree
         /// The tree being edited.
         /// </summary>
         BTree _tree;
+
+        public NodeView RootNodeView => FindNodeView(_tree.Root);
 
         public BTreeView()
         {
@@ -35,8 +38,53 @@ namespace BehaviorTree
             _tree = tree;
             graphViewChanged -= OnGraphViewChanged;
             DeleteElements(graphElements);
+
+            if (_tree == null) return;
             graphViewChanged += OnGraphViewChanged;
-            _tree.AllNodes.ForEach(CreateNodeView);
+
+            if (_tree.Root == null)
+            {
+                _tree.Root = ScriptableObject.CreateInstance<RootNode>();
+                _tree.Root.name = "Root";
+                _tree.Root.hideFlags = HideFlags.HideInHierarchy;
+                AssetDatabase.AddObjectToAsset(_tree.Root, _tree);
+                AssetDatabase.SaveAssets();
+                AssetDatabase.Refresh();
+            }
+            var rootAndNodes = _tree.AllNodes.Prepend(_tree.Root);
+
+            // Populate the graph view with its nodes.
+            foreach (var node in rootAndNodes)
+            {
+                CreateNodeView(node);
+            }
+
+            // Add the connections between the nodes.
+            foreach (var node in rootAndNodes)
+            {
+                var parentView = FindNodeView(node);
+                Node[] children = parentView.GetChildren();
+                for (int i = 0; i < children.Length; i++)
+                {
+                    Node child = children[i];
+                    var childView = FindNodeView(child);
+                    Edge edge = parentView.OutputPort.ConnectTo(childView.InputPort);
+                    AddElement(edge);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Returns the view of the given node, or null if it doesn't exist.
+        /// </summary>
+        private NodeView FindNodeView(Node node) => graphElements.ToList().OfType<NodeView>().FirstOrDefault(view => view.Node == node);
+
+        /// <summary>
+        /// Called when the user tries to create a new connection between two ports.
+        /// </summary>
+        public override List<Port> GetCompatiblePorts(Port startPort, NodeAdapter nodeAdapter)
+        {
+            return ports.ToList().Where(port => port.direction != startPort.direction && port.node != startPort.node).ToList();
         }
 
         /// <summary>
@@ -48,12 +96,35 @@ namespace BehaviorTree
             {
                 foreach (var element in graphViewChange.elementsToRemove)
                 {
+                    // Delete the node from the tree.
                     if (element is NodeView nodeView)
                     {
                         _tree.DeleteNode(nodeView.Node);
                     }
+
+                    // Delete a connection between two nodes.
+                    else if (element is Edge edge)
+                    {
+                        if (edge.input.node is NodeView childView && edge.output.node is NodeView parentView)
+                        {
+                            parentView.RemoveChild(childView);
+                        }
+                    }
                 }
             }
+
+            if (graphViewChange.edgesToCreate != null)
+            {
+                foreach (var edge in graphViewChange.edgesToCreate)
+                {
+                    // Create a connection between two nodes.
+                    if (edge.input.node is NodeView child && edge.output.node is NodeView parent)
+                    {
+                        parent.AddChild(child);
+                    }
+                }
+            }
+
             return graphViewChange;
         }
 
@@ -71,7 +142,9 @@ namespace BehaviorTree
             // If the user right clicks on a node, show the menu to delete it.
             if (evt.target is NodeView nodeView)
             {
-                evt.menu.AppendAction("Delete", (a) => DeleteElements(new[] { nodeView }));
+                if (nodeView.Node != _tree.Root)
+                    evt.menu.AppendAction("Delete", (a) => DeleteElements(new[] { nodeView }));
+
                 return;
             }
 
@@ -111,9 +184,12 @@ namespace BehaviorTree
         {
             if (selection.Count == 0) return;
             var nodes = selection.OfType<NodeView>();
+            ClearSelection();
+
             foreach (var node in nodes)
             {
                 CreateNode(node.Node.GetType(), node.GetPosition().position + Vector2.right * 20);
+                selection.Add(graphElements.Last());
             }
         }
     }
